toy_diffusion_rl/requirements.txt
# Toy Diffusion RL Requirements
# Core dependencies
torch>=2.0.0
numpy>=1.21.0
gymnasium>=0.29.0

# Configuration
pyyaml>=6.0

# Visualization
matplotlib>=3.5.0
tqdm>=4.64.0

# Optional: TensorBoard logging
# tensorboard>=2.10.0

# Optional: For development
# pytest>=7.0.0
# black>=22.0.0
# isort>=5.0.0
